requirements.txt
dataframe_image==0.2.2
pathlib==1.0.1
matplotlib==3.5.2
nltk==3.7
numpy==1.21.5
pandas==1.4.4
Pillow==9.2.0
Pillow==10.0.1
scikit_learn==1.0.2
scipy==1.9.1
seaborn==0.11.2
statsmodels==0.13.2
streamlit==1.16.0
streamlit==1.27.2
wordcloud==1.9.2
xgboost==1.7.4
